モデルの型をintからdoubleに変更

+++ minden/lib/features/power_plant/data/model/power_plant_detail_model.dart
@@ -9,7 +9,7 @@ class PowerPlantDetailModel extends PowerPlantDetail {
     required String voltageType,
     required String powerGenerationMethod,
     required String renewableType,
-    required int generationCapacity,
+    required double generationCapacity,
     required int displayOrder,
     required bool isRecommend,
     required String ownerName,
